+++ src/rust/scheduler/src/lib.rs
@@ -1,3 +1,6 @@
+pub mod slurm;
+use slurm::PartitionTable;
+
 use extendr_api::Result;
 use extendr_api::prelude::*;
 
@@ -7,8 +10,9 @@ use which::which;
 // pharos scheduler crate
 use nonmem::{RunOptions, expand_model_pattern};
 use scheduler::{
-    SchedulerType, sge::SubmitOptions as SgeSubmitOptions,
-    slurm::SubmitOptions as SlurmSubmitOptions,
+    SchedulerType,
+    sge::SubmitOptions as SgeSubmitOptions,
+    slurm::{SubmitOptions as SlurmSubmitOptions, resolve_partition},
 };
 
 use hyperion_core::{ResultExt, extendr_err};
@@ -120,6 +124,29 @@ pub fn submit_model_to_slurm(
     // Process model input to get list of model files
     let model_files = process_model_robj(model)?;
 
+    let (config_path, nonmem_config) = load_nonmem_config(None)?;
+
+    // check partition and give advice if needed
+    let model_count = model_files.len();
+    let ncpu_i32 = i32::from(ncpu.unwrap_or(1));
+    let table = PartitionTable::from_slurm()?;
+    let partition_name = resolve_partition(
+        partition.as_deref(),
+        nonmem_config.slurm.partition.as_deref(),
+    )
+    .map_to_extendr_err("Failed to get requested partition")?;
+    let active = table.find_partition(&partition_name);
+
+    if let Some(active) = active {
+        if ncpu_i32 > active.cpus as i32 {
+            let advice = table.partition_advice(ncpu_i32, &partition_name, model_count, false);
+            call!("stop", advice)?;
+        } else if table.is_underutilized(&partition_name, ncpu_i32, model_count) {
+            let advice = table.partition_advice(ncpu_i32, &partition_name, model_count, true);
+            call!("warning", advice)?;
+        }
+    }
+
     let submit_options = SlurmSubmitOptions {
         // process_model_robj is handling model paths so SubmitOptions doesn't need it.
         model: String::new(),
@@ -131,7 +158,6 @@ pub fn submit_model_to_slurm(
     };
 
     let scheduler = SchedulerType::new_slurm(submit_options);
-    let (config_path, nonmem_config) = load_nonmem_config(None)?;
     let parallel = ncpu.is_some_and(|n| n > 1);
 
     let run_options = RunOptions {
@@ -265,6 +291,8 @@ pub fn submit_model_to_sge(
 
 extendr_module! {
     mod hyperion_scheduler;
+
+    use slurm;
 
     fn submit_model_to_slurm;
     fn submit_model_to_sge;